Initial file check in
Ancient. CONVERTS A WEB PAGE HTML FILE INTO ASCII TEXT -- With source C code. This was an early version of an HTML stripper, it would remove known HTML tags which were common way back when. It was used with BBS systems before the DARPANet became the Internet which then became open to public use.